feat: add auth page
