+++ .prettierrc
@@ -1,4 +1,1 @@
-{
-  "singleAttributePerLine": true,
-  "plugins": ["prettier-plugin-tailwindcss"]
-}
+{ "singleAttributePerLine": true, "plugins": ["prettier-plugin-tailwindcss"] }
